requirement.txt
scrapy==1.6
Pillow==6.1.0
pyocr==0.7.2